Split calendar sharing security it test into company and person
